+++ src/main/java/plus/dragons/createenchantmentindustry/common/processing/forger/BlazeForgerBlockEntity.java
@@ -251,6 +251,14 @@ public class BlazeForgerBlockEntity extends BlazeExperienceBlockEntity {
                             component -> CEILang.builder().add(component).forGoggles(tooltip, 2),
                             TooltipFlag.NORMAL);
             }
+        } else {
+            if(inventory.forgingCompleted())
+                CEILang.translate("gui.goggles.forging.completed").style(ChatFormatting.GREEN).forGoggles(tooltip);
+            else if(!inventory.notEnoughItemToForge()){
+                if(inventory.incompatibleEnchantingTemplateType())
+                    CEILang.translate("gui.goggles.forging.invalid_template_type." + (special ? "normal": "special")).style(ChatFormatting.RED).forGoggles(tooltip);
+                else CEILang.translate("gui.goggles.forging.invalid_items").style(ChatFormatting.RED).forGoggles(tooltip);
+            }
         }
         return added;
     }